test: Test filters for non-processing
