+++ tests/integration/test_store.py
@@ -59,85 +59,6 @@ def test_legacy_store(mini_sentry, relay_chain):
     assert event["logentry"] == {"formatted": "Hello, World!"}
 
 
-@pytest.mark.parametrize(
-    "filter_config, should_filter",
-    [
-        ({"errorMessages": {"patterns": ["Panic: originalCreateNotification"]}}, True),
-        ({"errorMessages": {"patterns": ["Warning"]}}, False),
-    ],
-    ids=["error messages filtered", "error messages not filtered",],
-)
-def test_filters_are_applied(
-    mini_sentry, relay_with_processing, events_consumer, filter_config, should_filter,
-):
-    """
-    Test that relay normalizes messages when processing is enabled and sends them via Kafka queues
-    """
-    events_consumer = events_consumer()
-
-    relay = relay_with_processing()
-    project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    filter_settings = project_config["config"]["filterSettings"]
-    for key in filter_config.keys():
-        filter_settings[key] = filter_config[key]
-
-    # create a unique message so we can make sure we don't test with stale data
-    now = datetime.datetime.utcnow()
-    message_text = "some message {}".format(now.isoformat())
-
-    event = {
-        "message": message_text,
-        "exception": {
-            "values": [{"type": "Panic", "value": "originalCreateNotification"}]
-        },
-    }
-
-    relay.send_event(project_id, event)
-
-    if should_filter:
-        events_consumer.assert_empty()
-    else:
-        events_consumer.get_event()
-
-
-@pytest.mark.parametrize(
-    "is_enabled, should_filter",
-    [(True, True), (False, False),],
-    ids=["web crawlers filtered", "web crawlers not filtered",],
-)
-def test_web_crawlers_filter_are_applied(
-    mini_sentry, relay_with_processing, events_consumer, is_enabled, should_filter,
-):
-    """
-    Test that relay normalizes messages when processing is enabled and sends them via Kafka queues
-    """
-    relay = relay_with_processing()
-    project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    filter_settings = project_config["config"]["filterSettings"]
-    filter_settings["webCrawlers"] = {"isEnabled": is_enabled}
-
-    # UA parsing introduces higher latency in debug mode
-    events_consumer = events_consumer(timeout=10)
-
-    # create a unique message so we can make sure we don't test with stale data
-    now = datetime.datetime.utcnow()
-    message_text = "some message {}".format(now.isoformat())
-
-    event = {
-        "message": message_text,
-        "request": {"headers": {"User-Agent": "BingBot",}},
-    }
-
-    relay.send_event(project_id, event)
-
-    if should_filter:
-        events_consumer.assert_empty()
-    else:
-        events_consumer.get_event()
-
-
 @pytest.mark.parametrize("method_to_test", [("GET", False), ("POST", True)])
 def test_options_response(mini_sentry, relay, method_to_test):
     method, should_succeed = method_to_test
